Patient Chatbot/Backend/requirements.txt
requests~=2.31.0
flask~=2.3.0
flask-cors~=4.0.0
pandas~=2.2.0
openai==0.28.1
python-dotenv~=1.0.0
